# ddtrace/tracer: only finish execution trace task, restore pprof labels once