--- a/ddtrace/tracer/span.go
+++ b/ddtrace/tracer/span.go
@@ -478,9 +478,7 @@ func (s *span) Finish(opts ...ddtrace.FinishOption) {
 			s.Unlock()
 		}
 	}
-	if s.taskEnd != nil {
-		s.taskEnd()
-	}
+
 	if s.goExecTraced && rt.IsEnabled() {
 		// Only tag spans as traced if they both started & ended with
 		// execution tracing enabled. This is technically not sufficient
@@ -505,12 +503,6 @@ func (s *span) Finish(opts ...ddtrace.FinishOption) {
 	}
 
 	s.finish(t)
-
-	if s.pprofCtxRestore != nil {
-		// Restore the labels of the parent span so any CPU samples after this
-		// point are attributed correctly.
-		pprof.SetGoroutineLabels(s.pprofCtxRestore)
-	}
 }
 
 // SetOperationName sets or changes the operation name.
@@ -542,6 +534,9 @@ func (s *span) finish(finishTime int64) {
 	}
 	if s.Duration < 0 {
 		s.Duration = 0
+	}
+	if s.taskEnd != nil {
+		s.taskEnd()
 	}
 
 	keep := true
@@ -583,6 +578,12 @@ func (s *span) finish(finishTime int64) {
 			s, s.Name, s.Resource, s.Meta, s.Metrics)
 	}
 	s.context.finish()
+
+	if s.pprofCtxRestore != nil {
+		// Restore the labels of the parent span so any CPU samples after this
+		// point are attributed correctly.
+		pprof.SetGoroutineLabels(s.pprofCtxRestore)
+	}
 }
 
 // newAggregableSpan creates a new summary for the span s, within an application