--- a/internal/exectracetest/exectrace_go120.go
+++ b/internal/exectracetest/exectrace_go120.go
@@ -1,0 +1,8 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2024 Datadog, Inc.
+
+package exectracetest
+
+// Placeholder -- the latestgolang.org/x/exp does not support go1.20