@nsrip-dd 2024-05-24
For tracer reviewers, this is doing some cleanup under the span lock that wasn't under the lock before. The cleanup _should_ be cheap: just taking a stack trace with frame pointers, writing some bytes to a buffer, and updating a pointer in a goroutine structure. Our benchmarks don't show any regression, but I guess we also aren't specifically measuring contended access to a span while a program is calling `Finish` on it... Any thoughts on whether this will be a real performance issue?